Move MeteorStubs plugin up a directory
