+++ npm-packages/meteor-vite/src/meteor/IPC/methods/vite-server.ts
@@ -6,7 +6,7 @@ import Logger from '../../../utilities/Logger';
 import MeteorEvents, { MeteorIPCMessage } from '../MeteorEvents';
 import { MeteorViteConfig } from '../../../MeteorViteConfig';
 import { MeteorStubs } from '../../../plugin/internal';
-import { ProjectJson } from '../../../plugin/internal/MeteorStubs';
+import { ProjectJson } from '../../../plugin/MeteorStubs';
 import { RefreshNeeded } from '../../../ViteLoadRequest';
 import CreateIPCInterface, { IPCReply } from '../interface';
