+++ npm-packages/meteor-vite/src/plugin/internal/index.ts
@@ -1,1 +1,1 @@
-export { MeteorStubs } from './MeteorStubs';
+export { MeteorStubs } from '../MeteorStubs';
